+++ app/src/main/java/com/eddy/nrf/NewActivity.kt
@@ -6,272 +6,25 @@ import android.Manifest.permission.BLUETOOTH_ADVERTISE
 import android.Manifest.permission.BLUETOOTH_CONNECT
 import android.Manifest.permission.BLUETOOTH_SCAN
 import android.annotation.SuppressLint
-import android.bluetooth.BluetoothAdapter
-import android.bluetooth.BluetoothDevice
-import android.bluetooth.BluetoothGatt
-import android.bluetooth.BluetoothGattCharacteristic
-import android.bluetooth.BluetoothGattDescriptor
-import android.bluetooth.BluetoothGattServer
-import android.bluetooth.BluetoothGattServerCallback
-import android.bluetooth.BluetoothManager
-import android.bluetooth.BluetoothProfile
-import android.bluetooth.le.AdvertiseCallback
-import android.bluetooth.le.AdvertiseData
-import android.bluetooth.le.AdvertiseSettings
-import android.bluetooth.le.BluetoothLeAdvertiser
-import android.content.BroadcastReceiver
-import android.content.Context
-import android.content.Intent
-import android.content.IntentFilter
 import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
-import android.os.ParcelUuid
-import android.util.Log
 import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
-import com.eddy.nrf.Utils.byteArrayToHexArray
-import com.eddy.nrf.Utils.floatToByteArray
 import com.eddy.nrf.databinding.ActivityNewBinding
-import java.nio.ByteBuffer
-import java.util.Arrays
 
 
 private const val TAG = "NewActivity"
 
 class NewActivity : AppCompatActivity() {
+
     private lateinit var binding: ActivityNewBinding
-
-
-    /* Bluetooth API */
-    private lateinit var bluetoothManager: BluetoothManager
-    private var bluetoothGattServer: BluetoothGattServer? = null
-
-    /* Collection of notification subscribers */
-    private val registeredDevices = mutableSetOf<BluetoothDevice>()
-
-
-    private val heartRateNotificationHandler = android.os.Handler()
-
-    /**
-    - 거리 f4
-    - 속도 f4
-    - 파스? 기어 123  f1
-
-    다 줘야함
-
-    기어는 바꿀수도 있어야함
-     */
+    private lateinit var bluetoothServiceManager: BluetoothServiceManager
 
     var distance: Float = 15F //00000b67
     var speed: Float = 20F
     var gear = 1.toByte()
 
-
-
-    private val heartRateRunnable = object : Runnable {
-        override fun run() {
-
-            val buffer = ByteBuffer.allocate(9)
-            buffer.put(floatToByteArray(distance))
-            buffer.put(floatToByteArray(speed))
-            buffer.put(gear)
-            val resultArray = buffer.array()
-
-            Log.d(TAG, "거리+속도+기어 : ${byteArrayToHexArray(resultArray).joinToString(" ")}")
-
-            notifyHeartRate(resultArray)
-            heartRateNotificationHandler.postDelayed(this, 1000)
-        }
-    }
-
-
-    //리시버를 통해 notification 함
-    private val bluetoothReceiver = object : BroadcastReceiver() {
-        override fun onReceive(context: Context, intent: Intent) {
-            val state = intent.getIntExtra(BluetoothAdapter.EXTRA_STATE, BluetoothAdapter.STATE_OFF)
-
-            when (state) {
-                BluetoothAdapter.STATE_ON -> {
-                    startAdvertising()
-                    startServer()
-                    heartRateNotificationHandler.post(heartRateRunnable)  // Start heart rate notifications
-
-                }
-
-                BluetoothAdapter.STATE_OFF -> {
-                    stopServer()
-                    stopAdvertising()
-                    heartRateNotificationHandler.removeCallbacks(heartRateRunnable)  // Stop heart rate notifications
-
-                }
-            }
-        }
-    }
-
-    //광고에 대한 콜백
-    private val advertiseCallback = object : AdvertiseCallback() {
-        override fun onStartSuccess(settingsInEffect: AdvertiseSettings) {
-            Log.i(TAG, "LE Advertise Started.")
-        }
-
-        override fun onStartFailure(errorCode: Int) {
-            Log.w(TAG, "LE Advertise Failed: $errorCode")
-        }
-    }
-
-
-    //GATT에 대한 콜백
-    private val gattServerCallback = object : BluetoothGattServerCallback() {
-
-        override fun onConnectionStateChange(device: BluetoothDevice, status: Int, newState: Int) {
-            if (newState == BluetoothProfile.STATE_CONNECTED) {
-                Log.i(TAG, "BluetoothDevice CONNECTED: $device")
-            } else if (newState == BluetoothProfile.STATE_DISCONNECTED) {
-                Log.i(TAG, "BluetoothDevice DISCONNECTED: $device")
-                //Remove device from any active subscriptions
-                registeredDevices.remove(device)
-            }
-        }
-
-        @SuppressLint("MissingPermission")
-        override fun onCharacteristicReadRequest(
-            device: BluetoothDevice, requestId: Int, offset: Int,
-            characteristic: BluetoothGattCharacteristic
-        ) {
-            if (Utils.HEART_RATE_MEASUREMENT == characteristic.uuid) {
-                Log.i(TAG, "Read HeartRateMeasurement")
-                bluetoothGattServer?.sendResponse(
-                    device,
-                    requestId,
-                    BluetoothGatt.GATT_SUCCESS,
-                    0,
-                    byteArrayOf((60..100).random().toByte())
-                )
-            } else {
-                Log.w(TAG, "Invalid Characteristic Read: " + characteristic.uuid)
-                bluetoothGattServer?.sendResponse(
-                    device,
-                    requestId,
-                    BluetoothGatt.GATT_FAILURE,
-                    0,
-                    null
-                )
-            }
-        }
-
-        @SuppressLint("MissingPermission")
-        override fun onCharacteristicWriteRequest(
-            device: BluetoothDevice?,
-            requestId: Int,
-            characteristic: BluetoothGattCharacteristic?,
-            preparedWrite: Boolean,
-            responseNeeded: Boolean,
-            offset: Int,
-            value: ByteArray?
-        ) {
-            super.onCharacteristicWriteRequest(
-                device,
-                requestId,
-                characteristic,
-                preparedWrite,
-                responseNeeded,
-                offset,
-                value
-            )
-
-            Log.d(TAG, Arrays.toString(value))
-
-            val resultArray = value?.let { byteArrayToHexArray(it) }
-            gear = resultArray?.get(0)?.toByte() ?: 1
-
-            runOnUiThread {
-                binding.tvData.text = "speed: ${speed}\n distance: ${distance}\ngear: $gear"
-            }
-
-            if (responseNeeded) {
-                bluetoothGattServer?.sendResponse(
-                    device,
-                    requestId,
-                    BluetoothGatt.GATT_SUCCESS,
-                    0,
-                    null
-                )
-            }
-        }
-
-        //Descriptorr CCCD를 써야 노티를 보낼 수 있음
-        @SuppressLint("MissingPermission")
-        override fun onDescriptorReadRequest(
-            device: BluetoothDevice, requestId: Int, offset: Int,
-            descriptor: BluetoothGattDescriptor
-        ) {
-            if (Utils.CLIENT_CONFIG == descriptor.uuid) {
-                Log.d(TAG, "Config descriptor read")
-                val returnValue = if (registeredDevices.contains(device)) {
-                    BluetoothGattDescriptor.ENABLE_NOTIFICATION_VALUE
-                } else {
-                    BluetoothGattDescriptor.DISABLE_NOTIFICATION_VALUE
-                }
-                bluetoothGattServer?.sendResponse(
-                    device,
-                    requestId,
-                    BluetoothGatt.GATT_SUCCESS,
-                    0,
-                    returnValue
-                )
-            } else {
-                Log.w(TAG, "Unknown descriptor read request")
-                bluetoothGattServer?.sendResponse(
-                    device,
-                    requestId,
-                    BluetoothGatt.GATT_FAILURE,
-                    0, null
-                )
-            }
-        }
-
-        @SuppressLint("MissingPermission")
-        override fun onDescriptorWriteRequest(
-            device: BluetoothDevice, requestId: Int,
-            descriptor: BluetoothGattDescriptor,
-            preparedWrite: Boolean, responseNeeded: Boolean,
-            offset: Int, value: ByteArray
-        ) {
-            if (Utils.CLIENT_CONFIG == descriptor.uuid) {
-                if (Arrays.equals(BluetoothGattDescriptor.ENABLE_NOTIFICATION_VALUE, value)) {
-                    Log.d(TAG, "Subscribe device to notifications: $device")
-                    registeredDevices.add(device)
-                } else if (Arrays.equals(
-                        BluetoothGattDescriptor.DISABLE_NOTIFICATION_VALUE,
-                        value
-                    )
-                ) {
-                    Log.d(TAG, "Unsubscribe device from notifications: $device")
-                    registeredDevices.remove(device)
-                }
-
-                if (responseNeeded) {
-                    bluetoothGattServer?.sendResponse(
-                        device,
-                        requestId,
-                        BluetoothGatt.GATT_SUCCESS,
-                        0, null
-                    )
-                }
-            } else {
-                Log.w(TAG, "Unknown descriptor write request")
-                if (responseNeeded) {
-                    bluetoothGattServer?.sendResponse(
-                        device,
-                        requestId,
-                        BluetoothGatt.GATT_FAILURE,
-                        0, null
-                    )
-                }
-            }
-        }
-    }
 
     @SuppressLint("MissingPermission")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -280,98 +33,31 @@ class NewActivity : AppCompatActivity() {
         binding = ActivityNewBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        checkPermission()
+
+        bluetoothServiceManager = BluetoothServiceManager(this)
+        bluetoothServiceManager.initializeBluetooth()
+
         binding.tvData.text = "speed: ${speed}\n distance: ${distance}\ngear: $gear"
 
         // Devices with a display should not go to sleep
         window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-
-        initialization()
-
-        bluetoothManager = getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager
-        val bluetoothAdapter = bluetoothManager.adapter
-
-
-        Log.d(TAG, "Bluetooth name: " + bluetoothAdapter.address)
-        Log.d(TAG, "Bonded devices: " + bluetoothAdapter.bondedDevices)
-        // We can't continue without proper Bluetooth support
-        if (!checkBluetoothSupport(bluetoothAdapter)) {
-            finish()
-        }
-
-        // Register for system Bluetooth events
-        val filter = IntentFilter(BluetoothAdapter.ACTION_STATE_CHANGED)
-        registerReceiver(bluetoothReceiver, filter)
-        if (!bluetoothAdapter.isEnabled) {
-            Log.d(TAG, "Bluetooth is currently disabled...enabling")
-            bluetoothAdapter.enable()
-        } else {
-            Log.d(TAG, "Bluetooth enabled...starting services")
-            startAdvertising()
-            startServer()
-            heartRateNotificationHandler.post(heartRateRunnable)  // Start heart rate notifications
-        }
     }
 
-    override fun onStart() {
-        super.onStart()
-        // Register for system clock events
-        val filter = IntentFilter().apply {
-            addAction(Intent.ACTION_TIME_TICK)
-            addAction(Intent.ACTION_TIME_CHANGED)
-            addAction(Intent.ACTION_TIMEZONE_CHANGED)
-        }
-
-//        registerReceiver(timeReceiver, filter)
-    }
-
-    @SuppressLint("MissingPermission")
-    private fun notifyHeartRate(heartRate: ByteArray) {
-        if (registeredDevices.isEmpty()) return
-
-        Log.i(TAG, "Sending heart rate update to ${registeredDevices.size} subscribers")
-        for (device in registeredDevices) {
-            val heartRateCharacteristic = bluetoothGattServer
-                ?.getService(Utils.HEART_RATE_SERVICE)
-                ?.getCharacteristic(Utils.HEART_RATE_MEASUREMENT)
-            heartRateCharacteristic?.value = heartRate
-
-            bluetoothGattServer?.notifyCharacteristicChanged(device, heartRateCharacteristic, false)
-        }
-    }
 
     override fun onStop() {
         super.onStop()
-//        unregisterReceiver(timeReceiver)
+
+        bluetoothServiceManager.cleanup()
     }
 
     override fun onDestroy() {
         super.onDestroy()
-        val bluetoothAdapter = bluetoothManager.adapter
-        if (bluetoothAdapter.isEnabled) {
-            stopServer()
-            stopAdvertising()
-        }//앱이 종료되면 GATT서버, 광고 다 종료
 
-        unregisterReceiver(bluetoothReceiver)
+        bluetoothServiceManager.cleanup()
     }
 
-
-    private fun checkBluetoothSupport(bluetoothAdapter: BluetoothAdapter?): Boolean {
-
-        if (bluetoothAdapter == null) {
-            Log.w(TAG, "Bluetooth is not supported")
-            return false
-        }
-
-        if (!packageManager.hasSystemFeature(PackageManager.FEATURE_BLUETOOTH_LE)) {
-            Log.w(TAG, "Bluetooth LE is not supported")
-            return false
-        }
-
-        return true
-    }
-    private fun initialization() {
-
+    private fun checkPermission() {
         // 런타임 권한 확인
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
             if (checkSelfPermission(ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED ||
@@ -387,7 +73,6 @@ class NewActivity : AppCompatActivity() {
                 requestBlePermissions()
                 return
             }
-
         }
     }
 
@@ -408,57 +93,4 @@ class NewActivity : AppCompatActivity() {
             )
         }
     }
-
-    //광고 시작
-    @SuppressLint("MissingPermission")
-    private fun startAdvertising() {
-        val bluetoothLeAdvertiser: BluetoothLeAdvertiser? =
-            bluetoothManager.adapter.bluetoothLeAdvertiser
-
-        bluetoothLeAdvertiser?.let {
-            val settings = AdvertiseSettings.Builder()
-                .setAdvertiseMode(AdvertiseSettings.ADVERTISE_MODE_BALANCED)
-                .setConnectable(true)
-                .setTimeout(0)
-                .setTxPowerLevel(AdvertiseSettings.ADVERTISE_TX_POWER_MEDIUM)
-                .build()
-
-            val data = AdvertiseData.Builder()
-                .setIncludeDeviceName(true)
-                .setIncludeTxPowerLevel(false)
-                .addServiceUuid(ParcelUuid(Utils.HEART_RATE_SERVICE))
-                .build()
-
-            it.startAdvertising(settings, data, advertiseCallback)
-        } ?: Log.w(TAG, "Failed to create advertiser")
-    }
-
-
-    //광고 종료
-    @SuppressLint("MissingPermission")
-    private fun stopAdvertising() {
-        val bluetoothLeAdvertiser: BluetoothLeAdvertiser? =
-            bluetoothManager.adapter.bluetoothLeAdvertiser
-        bluetoothLeAdvertiser?.let {
-            it.stopAdvertising(advertiseCallback)
-        } ?: Log.w(TAG, "Failed to create advertiser")
-    }
-
-    //GATT서버 인스턴스 생성. services/characteristics을 기반으로
-    @SuppressLint("MissingPermission")
-    private fun startServer() {
-        bluetoothGattServer = bluetoothManager.openGattServer(this, gattServerCallback)
-
-        bluetoothGattServer?.addService(Utils.createHeartRateService())
-            ?: Log.w(TAG, "Unable to create GATT server")
-
-    }
-
-
-    //GATT서버 종료
-    @SuppressLint("MissingPermission")
-    private fun stopServer() {
-        bluetoothGattServer?.close()
-    }
-
 }
